worker/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.7.1
pika==1.3.2
langgraph==0.1.9
langchain-core==0.2.22
langchain==0.2.10
langchain-google-community==1.0.7
beautifulsoup4==4.12.3

